refactor:change related to hook
